Move GetOIDCIssuer to oidcx package
This decouples common utilities from server-specific code.
