+++ pkg/oidcx/oidcx_suite_test.go
@@ -1,0 +1,13 @@
+package oidcx_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestOidcx(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Oidcx Suite")
+}
