+++ pkg/oidcx/package.go
@@ -1,0 +1,1 @@
+package oidcx // import "code.cloudfoundry.org/perm/pkg/oidcx"
